+++ apps/trainer/app/notification/_components/NotificationItemContainer.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { NotificationInfo } from "@5unwan/core/api/types/common";
 import { useQuery } from "@tanstack/react-query";
 import NotificationItem from "@ui/components/NotificationItem/NotificationItem";
@@ -6,36 +8,48 @@ import { MouseEventHandler } from "react";
 
 import { notificationQueries } from "@trainer/queries/notification";
 
-import NotificationItemError from "./NotificationItemError";
 import NotificationItemFallback from "./NotificationItemFallback";
 import { parseContent } from "../_utils/notificationParser";
 
-type NotificationItemContainerProps = {
+type Props = {
   notification: NotificationInfo;
   onClick: MouseEventHandler<HTMLLIElement>;
 };
 
-function NotificationItemContainer({ notification, onClick }: NotificationItemContainerProps) {
-  const { data, isPending, isError } = useQuery({
-    ...notificationQueries.detail(notification.notificationId),
-  });
+function NotificationItemContainer({ notification, onClick }: Props) {
+  const { data, isLoading, isError } = useQuery(
+    notificationQueries.detail(notification.notificationId),
+  );
 
   const { content, type, sendDate, isProcessed, notificationId } = notification;
   const { message, eventDate } = parseContent(content);
 
-  if (isPending)
+  if (isLoading || !data) {
     return (
       <NotificationItemFallback
+        message={message || ""}
+        eventDate={eventDate || ""}
         variant={type}
         createdAt={sendDate}
         isCompleted={isProcessed}
-        message={message || ""}
-        eventDate={eventDate || ""}
       />
     );
+  }
 
-  if (isError) return <NotificationItemError />;
-
+  if (isError) {
+    return (
+      <NotificationItem
+        message={message || ""}
+        eventDate={eventDate || ""}
+        variant={type}
+        createdAt={sendDate}
+        isCompleted={isProcessed}
+        key={`notification-${notificationId}`}
+        className="w-full"
+        onClick={onClick}
+      />
+    );
+  }
   const { profilePictureUrl, name } = data.data.userDetail;
 
   return (
